+++ lib/core/di/supabase_module.dart
